backend/requirements.txt
fastapi
uvicorn
ultralytics
opencv-python
numpy
pillow
easyocr
